requirements.txt
streamlit==1.28.1
opencv-python-headless==4.8.1.78
numpy==1.24.3
pandas==2.0.3
pillow==10.0.1
plotly==5.17.0
matplotlib==3.7.2
seaborn==0.12.2
